requirements.txt
alembic==0.8.2
amqp==1.4.6
anyjson==0.3.3
astroid==1.3.8
backports.ssl-match-hostname==3.4.0.2
bcrypt==2.0.0
billiard==3.3.0.20
celery==3.1.18
certifi==2015.9.6.2
cffi==1.2.1
coverage==4.0.2
decorator==4.0.4
Django==1.8.5
django-braces==1.8.1
django-celery==3.1.17
django-crispy-forms==1.5.2
django-extensions==1.5.7
django-mptt==0.7.4
django-nose==1.4.2
django-reversion==1.9.3
django-sniplates==0.3.2
Flask==0.10.1
Flask-Login==0.2.11
Flask-Migrate==1.5.1
Flask-Moment==0.5.1
Flask-Permissions==0.2.3
Flask-PyMongo==0.3.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
functools32==3.2.3.post2
idna==2.0
ipaddr==2.1.11
iptools==0.6.1
ipykernel==4.1.0
ipython==4.0.0
ipython-genutils==0.1.0
ipywidgets==4.0.3
itsdangerous==0.24
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.1.1
jupyter-console==4.0.3
jupyter-core==4.0.6
kombu==3.0.26
logilab-common==1.1.0
Mako==1.0.2
markup==0.2
MarkupSafe==0.23
mistune==0.7.1
MySQL-python==1.2.5
nbconvert==4.0.0
nbformat==4.0.1
nose==1.3.7
notebook==4.0.6
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
ptyprocess==0.5
py==1.4.30
py-xlsx==0.4.1
pycparser==2.14
pydot==1.0.2
Pygments==2.0.2
pygraphviz==1.3.1
pylint==1.4.4
pylint-django==0.6.1
pylint-plugin-utils==0.2.3
pymongo==2.8.1
PyMySQL==0.6.6
pyparsing==1.5.7
pytest==2.8.1
python-editor==0.4
python-libnmap==0.6.3
pytz==2015.4
pyzmq==14.7.0
qtconsole==4.0.1
simplegeneric==0.8.1
six==1.9.0
SQLAlchemy==1.0.8
SQLAlchemy-Utils==0.31.0
terminado==0.5
tornado==4.2.1
traitlets==4.0.0
Werkzeug==0.10.4
xlsx2csv==0.7.2
